requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.5
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
Pillow==9.0.0
prompt-toolkit==3.0.27
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
